Removed old libraryproperties and replaced it with tabbed one
