+++ src/main/java/org/jabref/gui/libraryproperties/AbstractPropertiesTabView.java
@@ -1,4 +1,4 @@
-package org.jabref.gui.newlibraryproperties;
+package org.jabref.gui.libraryproperties;
 
 import javax.inject.Inject;
