+++ src/main/java/org/jabref/gui/libraryproperties/LibraryProperties.fxml
@@ -4,7 +4,7 @@
 <?import javafx.scene.control.DialogPane?>
 <?import javafx.scene.control.TabPane?>
 <DialogPane xmlns="http://javafx.com/javafx" xmlns:fx="http://javafx.com/fxml"
-            fx:controller="org.jabref.gui.newlibraryproperties.LibraryPropertiesView"
+            fx:controller="org.jabref.gui.libraryproperties.LibraryPropertiesView"
             prefHeight="400.0" prefWidth="600.0">
     <content>
         <TabPane fx:id="tabPane" tabClosingPolicy="UNAVAILABLE"/>
