+++ src/main/java/org/jabref/gui/libraryproperties/LibraryPropertiesView.java
@@ -1,4 +1,4 @@
-package org.jabref.gui.newlibraryproperties;
+package org.jabref.gui.libraryproperties;
 
 import javax.inject.Inject;
