+++ src/main/java/org/jabref/gui/libraryproperties/PropertiesTabViewModel.java
@@ -1,4 +1,4 @@
-package org.jabref.gui.newlibraryproperties;
+package org.jabref.gui.libraryproperties;
 
 public interface PropertiesTabViewModel {
