+++ src/main/java/org/jabref/gui/libraryproperties/saving/SavingPropertiesView.java
@@ -1,4 +1,4 @@
-package org.jabref.gui.newlibraryproperties.saving;
+package org.jabref.gui.libraryproperties.saving;
 
 import javax.inject.Inject;
 
@@ -7,8 +7,8 @@ import javafx.scene.control.CheckBox;
 
 import org.jabref.gui.commonfxcontrols.FieldFormatterCleanupsPanel;
 import org.jabref.gui.commonfxcontrols.SaveOrderConfigPanel;
-import org.jabref.gui.newlibraryproperties.AbstractPropertiesTabView;
-import org.jabref.gui.newlibraryproperties.PropertiesTab;
+import org.jabref.gui.libraryproperties.AbstractPropertiesTabView;
+import org.jabref.gui.libraryproperties.PropertiesTab;
 import org.jabref.logic.l10n.Localization;
 import org.jabref.model.database.BibDatabaseContext;
 import org.jabref.preferences.PreferencesService;
